requirements.txt
Flask==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
idna==3.11
itsdangerous==2.2.0
nba_api>=1.4.1
pandas>=2.0.0
python-dotenv==1.2.2
requests==2.32.5
supabase>=2.0.0
urllib3==2.6.3
